Finish convert to wpf
