+++ DocumentComplianceChecker_HSEproject/App.xaml
@@ -2,6 +2,6 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              StartupUri="MainWindow.xaml">
-	<Application.Resources>
-	</Application.Resources>
+    <Application.Resources>
+    </Application.Resources>
 </Application>
