+++ DocumentComplianceChecker_HSEproject/DocumentComplianceChecker_HSEproject.csproj
@@ -6,16 +6,9 @@
 		<Nullable>enable</Nullable>
 		<ImplicitUsings>enable</ImplicitUsings>
 	</PropertyGroup>
-	<ItemGroup>
-	  <None Remove="App.xml" />
-	</ItemGroup>
-	<ItemGroup>
-	  <ApplicationDefinition Include="App.xml" />
-	</ItemGroup>
 
-  <ItemGroup>
-    <PackageReference Include="DocumentFormat.OpenXml" Version="3.3.0" />
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="9.0.4" />
-  </ItemGroup>
-
+	<ItemGroup>
+		<PackageReference Include="DocumentFormat.OpenXml" Version="3.1.0" />
+		<PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="8.0.0" />
+	</ItemGroup>
 </Project>
