+++ DocumentComplianceChecker_HSEproject/MainWindow.xaml
@@ -1,12 +1,26 @@
 ﻿<Window x:Class="DocumentComplianceChecker_HSEproject.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:DocumentComplianceChecker_HSEproject"
-        mc:Ignorable="d"
-        Title="MainWindow" Height="450" Width="800">
+        Title="Document Compliance Checker" Height="400" Width="600">
     <Grid>
-        
+        <Grid.RowDefinitions>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition Height="*"/>
+        </Grid.RowDefinitions>
+
+        <!-- Панель управления -->
+        <StackPanel Grid.Row="0" Orientation="Vertical" Margin="5">
+            <TextBlock Text="Выберите режим проверки:" FontSize="14" Margin="0,0,0,5"/>
+            <StackPanel Orientation="Horizontal" Margin="0,0,0,5">
+                <RadioButton x:Name="StandardRulesRadio" Content="Стандартные правила" IsChecked="True" Margin="0,0,10,0"/>
+                <RadioButton x:Name="TemplateRadio" Content="Использовать шаблон" />
+            </StackPanel>
+            <Button Content="Проверить документ" Width="150" Height="30" Click="CheckDocument_Click"/>
+        </StackPanel>
+
+        <!-- Лог -->
+        <ScrollViewer Grid.Row="1" VerticalScrollBarVisibility="Auto" Margin="5">
+            <TextBlock x:Name="LogTextBlock" TextWrapping="Wrap"/>
+        </ScrollViewer>
     </Grid>
 </Window>
