+++ DocumentComplianceChecker_HSEproject/Services/FileManager.cs
@@ -1,4 +1,5 @@
 ﻿using DocumentComplianceChecker_HSEproject.Interfaces;
+using System.IO;
 
 namespace DocumentComplianceChecker_HSEproject.Services
 {
